==> 1.hourly_analysis/2.scripts/3.matching_analysis/1.experiments_cruise/2.experiment_exit_cruise/3_script_checking_intervention.Rmd <==
---
title: "Checking the Hypothetical Intervention - Exiting Cruise Experiment"
description: |
  Comparing hours with exiting cruise traffic to hours without. Adjusting for calendar and weather indicators.
author:
  - name: Marie-Abèle Bind 
    url: https://scholar.harvard.edu/marie-abele
    affiliation: Biostatistics Center, Massachusetts General Hospital
    affiliation_url: https://biostatistics.massgeneral.org/faculty/marie-abele-bind-phd/
  - name: Marion Leroutier 
    url: https://www.parisschoolofeconomics.eu/en/leroutier-marion/work-in-progress/
    affiliation: Misum, Stockholm School of Economics
    affiliation_url: https://www.hhs.se/en/persons/l/leroutier-marion/
  - name: Léo Zabrocki 
    url: https://lzabrocki.github.io/
    affiliation: Paris School of Economics
    affiliation_url: https://www.parisschoolofeconomics.eu/fr/zabrocki-leo/
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
      keep_md: true
      toc: true
      toc_depth: 2
---

<style>
body {
text-align: justify}
</style>

In this document, we take great care providing all steps and R codes required to check the intervention we set up in our matching procedure. We compare hours where:

* treated units are hours with positive exiting cruise traffic in t.
* control units are hours without exiting cruise traffic in t.

We adjust for calendar calendar indicator and weather confouding factors.

**Should you have any questions, need help to reproduce the analysis or find coding errors, please do not hesitate to contact us at leo.zabrocki@psemail.eu and marion.leroutier@psemail.eu.**

# Required Packages

To reproduce exactly the `3_script_checking_intervention.html` document, we first need to have installed:

* the [R](https://www.r-project.org/) programming language 
* [RStudio](https://rstudio.com/), an integrated development environment for R, which will allow you to knit the `3_script_checking_intervention.Rmd` file and interact with the R code chunks
* the [R Markdown](https://rmarkdown.rstudio.com/) package
* and the [Distill](https://rstudio.github.io/distill/) package which provides the template for this document. 

Once everything is set up, we have to load the following packages:

```{r, echo = TRUE, message = FALSE, warning = FALSE}
# load required packages
library(knitr) # for creating the R Markdown document
library(here) # for files paths organization
library(tidyverse) # for data manipulation and visualization
library(ggridges) # for ridge density plots
library(Cairo) # for printing customed police of graphs
library(patchwork) # combining plots
```

We finally load our customed `ggplot2` theme for graphs:

```{r, echo = TRUE, message = FALSE, warning = FALSE}
# load ggplot customed theme
source(
  here::here(
    "2.scripts",
    "4.custom_ggplot2_theme",
    "script_custom_ggplot_theme.R"
  )
)
````

The theme is based on the fantastic [hrbrthemes](https://hrbrmstr.github.io/hrbrthemes/index.html) package. If you do not want to use this theme or are unable to install it because of fonts issues, you can use the `theme_bw()` already included in the `ggplot2` package.

# Preparing the Data

We load the matched data:

```{R, echo = TRUE}
# load matched data
data_matched <-
  readRDS(
    here::here(
      "1.data",
      "2.data_for_analysis",
      "1.matched_data",
      "1.experiments_cruise",
      "2.experiment_exit_cruise",
      "matched_data_exit_cruise.rds"
    )
  )
```

# Checking the Hypothetical Intervention

We compute the difference in the cruise entering total gross tonnage for each pair:

```{R, echo=TRUE, message = FALSE, warning = FALSE}
# compute the difference in tonnage by pair
pair_difference_tonnage_t <- data_matched %>%
  select(total_gross_tonnage_exit_cruise, is_treated, pair_number) %>%
  arrange(pair_number, is_treated) %>%
  select(-is_treated) %>%
  group_by(pair_number) %>%
  summarise(tonnage_difference = total_gross_tonnage_exit_cruise[2] - total_gross_tonnage_exit_cruise[1])
````

We find on average, a `r round(mean(pair_difference_tonnage_t$tonnage_difference),0)` difference in gross tonnage between treated and control units. Below is the distribution of the pair difference in hourly gross tonnage in t:

```{R, echo=TRUE, message = FALSE, warning = FALSE, layout="l-body-outset", fig.width=14, fig.height=10, dev = "CairoPNG"}
# plot the graph
graph_tonnage_difference_density <-
  ggplot(pair_difference_tonnage_t, aes(x = tonnage_difference)) +
  geom_density(
    colour = "black",
    fill = "deepskyblue3",
    size = 1.1,
    alpha = 0.8
  ) +
  geom_vline(
    xintercept = mean(pair_difference_tonnage_t$tonnage_difference),
    size = 1.1,
    color = "coral") +
  scale_x_continuous(
    breaks = scales::pretty_breaks(n = 8),
    labels = function(x)
      format(x, big.mark = " ", scientific = FALSE)
  ) +
  xlab("Pair Difference in Tonnage") + ylab("Density") +
  custom_theme

# we print the graph
graph_tonnage_difference_density

ggsave(
  graph_tonnage_difference_density,
  filename = here::here(
    "3.outputs",
    "1.figures",
    "2.experiments_cruise",
    "2.experiment_exit_cruise",
    "1.checking_matching_procedure",
    "graph_tonnage_difference_exiting_cruise_tonnage_density.pdf"
  ),
  width = 30,
  height = 20,
  units = "cm",
  device = cairo_pdf
)
````

To check which hypothetical intervention we study, we plot below the average tonnage for each hour and for treated and control groups :

```{R, echo=TRUE, message = FALSE, warning = FALSE, layout="l-body-outset", fig.width=20, fig.height=10, dev = "CairoPNG"}
# compute mean tonnage for each day
mean_tonnage_group_hour <- data_matched %>%
  # select relevant variables
  select(pair_number,
         is_treated,
         contains("total_gross_tonnage_exit_cruise")) %>%
  # transform data in long format
  pivot_longer(
    cols = -c(pair_number, is_treated),
    names_to = "variable",
    values_to = "tonnage"
  ) %>%
  # create the day variable
  mutate(
    time = 0 %>%
      ifelse(str_detect(variable, "lag_1"),-1, .) %>%
      ifelse(str_detect(variable, "lag_2"),-2, .) %>%
      ifelse(str_detect(variable, "lag_3"),-3, .) %>%
      ifelse(str_detect(variable, "lead_1"), 1, .) %>%
      ifelse(str_detect(variable, "lead_2"), 2, .) %>%
      ifelse(str_detect(variable, "lead_3"), 3, .)
  ) %>%
  # rename the labels of the is_treated dummy
  mutate(is_treated = ifelse(is_treated == TRUE, "Treated", "Control")) %>%
  # compute the mean tonnage for each day and pollutant
  group_by(variable, is_treated, time) %>%
  summarise(tonnage = mean(tonnage))

# plot the graph
graph_mean_tonnage_group_hour <-
  ggplot(
    mean_tonnage_group_hour,
    aes(
      x = as.factor(time),
      y = tonnage,
      group = is_treated,
      colour = is_treated,
      fill = is_treated
    )
  )  +
  geom_segment(
    x = 4,
    y = 0,
    xend = 4,
    yend = 63722,
    lineend = "round",
    # See available arrow types in example above
    linejoin = "round",
    size = 0.5,
    colour = "black"
  ) +
  geom_line() +
  geom_point(shape = 21,
             size = 8,
             colour = "black") +
  scale_colour_manual(values = c("deepskyblue3", "tomato")) +
  scale_fill_manual(values = c("deepskyblue3", "tomato")) +
  scale_y_continuous(
    breaks = scales::pretty_breaks(n = 5),
    labels = function(x)
      format(x, big.mark = " ", scientific = FALSE)
  ) +
  labs(fill = "Group:") +
  xlab("Hour") + ylab("Hourly Total Gross Tonnage") +
  custom_theme +
  guides(color = FALSE) +
  theme(
    # axis titles parameters
    axis.title.x = element_text(size=40, face = "bold", margin = margin(t = 20, r = 0, b = 0, l =0)),
    axis.title.y = element_text(size=40, face = "bold", margin = margin(t = 0, r = 20, b = 0, l = 0)),
    # axis texts
    axis.text.x = element_text(size=30),
    axis.text.y = element_text(size=30),
    # facet texts
    strip.text.x = element_text(size=40, face = "bold"),
    strip.text.y = element_text(size=40, face = "bold"),
    # legend parameters
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal"
  )

# we print the graph
graph_mean_tonnage_group_hour

# save the graph
ggsave(
  graph_mean_tonnage_group_hour,
  filename = here::here(
    "3.outputs",
    "1.figures",
    "2.experiments_cruise",
    "2.experiment_exit_cruise",
    "1.checking_matching_procedure",
    "graph_mean_tonnage_group_hour_exiting_cruise.pdf"
  ),
  width = 40,
  height = 20,
  units = "cm",
  device = cairo_pdf
)
````


# Checking Other Vessels' Types Traffic Evolution

We also check how the difference in tonnage for other vessels' types between treated and control units evolves:

```{R, echo=TRUE, message = FALSE, warning = FALSE, layout="l-body-outset", fig.width=25, fig.height=18, dev = "CairoPNG"}
# we create a table with the tonnage for each pair,
# for each vessel type,
# and for -6 hours to + 6 hours
data_vessel_type_tonnage <- data_matched %>%
  # relabel treatment indicator
  mutate(is_treated = ifelse(is_treated == TRUE, "treated", "control")) %>%
  # select relevant variables
  select(
    pair_number,
    is_treated,
    contains("total_gross_tonnage_entry_cruise"),
    contains("total_gross_tonnage_exit_cruise"),
    contains("total_gross_tonnage_entry_ferry"),
    contains("total_gross_tonnage_exit_ferry"),
    contains("total_gross_tonnage_entry_other_vessels"),
    contains("total_gross_tonnage_exit_other_vessels")
  ) %>%
  # transform data in long format
  pivot_longer(
    cols = -c(pair_number, is_treated),
    names_to = "variable",
    values_to = "tonnage"
  ) %>%
  # create vessel type variable
  mutate(
    vessel_type = NA %>%
      ifelse(str_detect(variable, "cruise"), "Cruise", .) %>%
      ifelse(str_detect(variable, "ferry"), "Ferry", .) %>%
      ifelse(str_detect(variable, "other_vessels"), "Other Vessels", .)
  ) %>%
  # create the day variable
  mutate(entry_exit = NA %>%
           ifelse(str_detect(variable, "entry"), "Arrival", .) %>%
           ifelse(str_detect(variable, "exit"), "Departure", .)) %>%
  mutate(
    time = 0 %>%
      ifelse(str_detect(variable, "lag_1"),-1, .) %>%
      ifelse(str_detect(variable, "lag_2"),-2, .) %>%
      ifelse(str_detect(variable, "lag_3"),-3, .) %>%
      ifelse(str_detect(variable, "lead_1"), 1, .) %>%
      ifelse(str_detect(variable, "lead_2"), 2, .) %>%
      ifelse(str_detect(variable, "lead_3"), 3, .)
  ) %>%
  select(pair_number, vessel_type, is_treated, entry_exit, time, tonnage) %>%
  pivot_wider(names_from = is_treated, values_from = tonnage)

# compute the average difference in traffic between treated and control units
data_mean_difference <- data_vessel_type_tonnage %>%
  mutate(difference = treated - control) %>%
  select(-c(treated, control)) %>%
  group_by(vessel_type, entry_exit, time) %>%
  summarise(mean_difference = mean(difference)) %>%
  ungroup()

# plot the evolution
graph_tonnage_difference_vessel_type <-
  ggplot(data_mean_difference,
         aes(x = as.factor(time), y = mean_difference, group = "l"))  +
  geom_hline(yintercept = 0) +
  geom_segment(aes(
    x = as.factor(time),
    xend = as.factor(time),
    y = 0,
    yend = mean_difference
  )) +
  geom_point(
    shape = 21,
    size = 8,
    colour = "black",
    fill = "deepskyblue3"
  ) +
  scale_y_continuous(
    breaks = scales::pretty_breaks(n = 5),
    labels = function(x)
      format(x, big.mark = " ", scientific = FALSE)
  ) +
  facet_grid(vessel_type ~ entry_exit) +
  xlab("Hour") + ylab("Average Difference in\n Total Gross Tonnage") +
  custom_theme +
  theme(
    # axis titles parameters
    axis.title.x = element_text(size=40, face = "bold", margin = margin(t = 20, r = 0, b = 0, l =0)),
    axis.title.y = element_text(size=40, face = "bold", margin = margin(t = 0, r = 20, b = 0, l = 0)),
    # axis texts
    axis.text.x = element_text(size=30),
    axis.text.y = element_text(size=30),
    # facet texts
    strip.text.x = element_text(size=40, face = "bold"),
    strip.text.y = element_text(size=40, face = "bold"),
    # legend parameters
    legend.position = "top",
    legend.justification = "left",
    legend.direction = "horizontal",
    legend.title = element_text(size=40, face = "bold"),
    legend.text = element_text(size=30)
  )

# we print the graph
graph_tonnage_difference_vessel_type

# save the graph
ggsave(
  graph_tonnage_difference_vessel_type,
  filename = here::here(
    "3.outputs",
    "1.figures",
    "2.experiments_cruise",
    "2.experiment_exit_cruise",
    "1.checking_matching_procedure",
    "graph_tonnage_difference_vessel_type.pdf"
  ),
  width = 50,
  height = 36,
  units = "cm",
  device = cairo_pdf
)
````
